[REFACTOR] 파일 위치 변경 및 사용하지 않는 파일 삭제
- templates/Dialog => modules/Dialog
- Menu 관련 dialog 삭제
