+++ src/components/modules/DeleteDialog/LectureRoomDeleteDialogContainer.js
@@ -6,9 +6,9 @@ import PropTypes from 'prop-types';
 import { useDispatch, useSelector } from 'react-redux';
 import DeleteDialog from './DeleteDialog';
 import {
-  removeSubject,
-  loadSubjects
-} from '../../../../modules/schedule/subject';
+  removeLectureRoom,
+  loadLectureRooms
+} from '../../../modules/schedule/lectureRoom';
 
 const DeleteDialogContainer = ({
   location,
@@ -19,16 +19,16 @@ const DeleteDialogContainer = ({
   type
 }) => {
   const dispatch = useDispatch();
-  const { remove, error } = useSelector(({ subject }) => ({
-    remove: subject.remove,
-    error: subject.removeError
+  const { remove, error } = useSelector(({ lectureRoom }) => ({
+    remove: lectureRoom.remove,
+    error: lectureRoom.removeError
   }));
 
   const onDelete = () => {
     const token = localStorage.getItem('token');
-    const id = data.subjectId;
-    console.log(data.subjectId);
-    dispatch(removeSubject({ id, token }));
+    const id = data.lectureRoomId;
+    console.log(data.lectureRoomId);
+    dispatch(removeLectureRoom({ id, token }));
   };
 
   useEffect(() => {
@@ -41,7 +41,7 @@ const DeleteDialogContainer = ({
         ignoreQueryPrefix: true
       });
       const token = localStorage.getItem('token');
-      dispatch(loadSubjects({ direction, page, size, token }));
+      dispatch(loadLectureRooms({ direction, page, size, token }));
       setOpen();
     }
   }, [remove, error, dispatch]);
